refactor [#55] utils 파일 이름 케밥케이스로 수정
